requirements.txt
omegaconf==2.3.0
python-dotenv==1.0.1
streamlit==1.32.2
streamlit_pills==0.3.0
streamlit-feedback==0.1.3
uuid==1.30
langdetect==1.0.9
langcodes==3.4.0
vectara-agentic==0.1.8
arize-phoenix==4.35.1
openinference-instrumentation-llama-index==3.0.2